Make fake clients

make_fake_clients() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake clients. The function returns Clients (a tibble).

Usage

make_fake_clients(
  datasets_ls = NULL,
  add_sports_1L_lgl = T,
  age_var_nm_1L_chr = "Age",
  annual_referrals_int = c(350, 500, 550),
  appointments_rows_1L_int = 30,
  athlete_roles_chr = c("Athlete", "AlumniAthlete"),
  burn_from_1L_chr = "2020-03-01",
  burn_referrals_int = c(200, 250),
  cancellations_probs_chr = c(0.05, 0.05, 0.03, 0.08, 0.1),
  cancellations_rows_1L_int = 247,
  categories_chr = make_sports_categories(),
  cleanse_ages_1L_lgl = T,
  clinician_dbl = c(0.15, 0.06, 0.01, 0.06, 0.68),
  clinicians_int = c(10, 5, 3, 4, 50),
  clinicians_severity_1L_int = 2,
  durations_chr = c("Under 30 mins", "Between 30 and 45 mins", "Between 45 and 75 mins",
    "50 mins"),
  end_date_1L_chr = "2024-03-30",
  neuropsychological_rows_1L_int = 1,
  prefix_1L_chr = "Client_",
  retain_1L_dbl = 0.3,
  scale_1L_dbl = 100,
  missing_1L_chr = character(0),
  moderate_int = c(2L, 4L),
  notes_rows_1L_int = 3L,
  para_probs_dbl = c(0.2, 0.3),
  path_1L_chr = character(0),
  referrals_cols_int = 4:8,
  retainer_rows_1L_int = 12,
  sessions_cuts_ls = list(c(1, 2), c(3, 5), c(6, 10), c(10, 15), c(16, 20), c(20, 25)),
  sessions_probs_ls = purrr::map(1:5, ~c(0.06, 0.26, 0.35, 0.17, 0.07, 0.09)),
  sessions_moderate_int = c(4, 15),
  share_at_weekend_dbl = 0.05,
  share_by_age_athletes_dbl = c(0.05, 0.25, 0.25, 0.2, 0.15, 0.1),
  share_by_age_other_dbl = c(0, 0.05, 0.1, 0.15, 0.15, 0.55),
  share_by_quarter_dbl = c(0.2, 0.25, 0.25, 0.3),
  share_by_reason_dbl = c(0.15, 0.06, 0.01, 0.06, 0.68),
  share_by_referrer_dbl = c(rep(0.1, 3), rep(0.06, 2), 0.1, 0.08, 0.1, 0.3),
  share_by_role_dbl = c(0.12, 0.65, 0.01, 0.2, 0.02),
  share_by_sex_dbl = c(0.05, 0.55, 0.4),
  share_is_para_1L_dbl = 0.28,
  sheets_int = c(1, 3:8),
  sports_1L_int = integer(0),
  sports_tab_1L_int = 2L,
  start_date_1L_chr = "2022-03-01",
  uid_vars_chr = c("Client ID", "UID"),
  unit_cost_lup = NULL
)

Arguments

datasets_ls

Datasets (a list), Default: NULL

add_sports_1L_lgl

Add sports (a logical vector of length one), Default: T

age_var_nm_1L_chr

Age variable name (a character vector of length one), Default: 'Age'

annual_referrals_int

Annual referrals (an integer vector), Default: c(350, 500, 550)

appointments_rows_1L_int

Appointments rows (an integer vector of length one), Default: 30

athlete_roles_chr

Athlete roles (a character vector), Default: c("Athlete", "AlumniAthlete")

burn_from_1L_chr

Burn from (a character vector of length one), Default: '2020-03-01'

burn_referrals_int

Burn referrals (an integer vector), Default: c(200, 250)

cancellations_probs_chr

Cancellations probabilities (a character vector), Default: c(0.05, 0.05, 0.03, 0.08, 0.1)

cancellations_rows_1L_int

Cancellations rows (an integer vector of length one), Default: 247

categories_chr

Categories (a character vector), Default: make_sports_categories()

cleanse_ages_1L_lgl

Cleanse ages (a logical vector of length one), Default: T

clinician_dbl

Clinician (a double vector), Default: c(0.15, 0.06, 0.01, 0.06, 0.68)

clinicians_int

Clinicians (an integer vector), Default: c(10, 5, 3, 4, 50)

clinicians_severity_1L_int

Clinicians severity (an integer vector of length one), Default: 2

durations_chr

Durations (a character vector), Default: c("Under 30 mins", "Between 30 and 45 mins", "Between 45 and 75 mins", "50 mins")

end_date_1L_chr

End date (a character vector of length one), Default: '2024-03-30'

neuropsychological_rows_1L_int

Neuropsychological rows (an integer vector of length one), Default: 1

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Client_'

retain_1L_dbl

Retain (a double vector of length one), Default: 0.3

scale_1L_dbl

Scale (a double vector of length one), Default: 100

missing_1L_chr

Missing (a character vector of length one), Default: character(0)

moderate_int

Moderate (an integer vector), Default: c(2L, 4L)

notes_rows_1L_int

Notes rows (an integer vector of length one), Default: 3

para_probs_dbl

Para probabilities (a double vector), Default: c(0.2, 0.3)

path_1L_chr

Path (a character vector of length one), Default: character(0)

referrals_cols_int

Referrals columns (an integer vector), Default: 4:8

retainer_rows_1L_int

Retainer rows (an integer vector of length one), Default: 12

sessions_cuts_ls

Sessions cuts (a list), Default: list(c(1, 2), c(3, 5), c(6, 10), c(10, 15), c(16, 20), c(20, 25))

sessions_probs_ls

Sessions probabilities (a list), Default: purrr::map(1:5, ~c(0.06, 0.26, 0.35, 0.17, 0.07, 0.09))

sessions_moderate_int

Sessions moderate (an integer vector), Default: c(4, 15)

share_at_weekend_dbl

Share at weekend (a double vector), Default: 0.05

share_by_age_athletes_dbl

Share by age athletes (a double vector), Default: c(0.05, 0.25, 0.25, 0.2, 0.15, 0.1)

share_by_age_other_dbl

Share by age other (a double vector), Default: c(0, 0.05, 0.1, 0.15, 0.15, 0.55)

share_by_quarter_dbl

Share by quarter (a double vector), Default: c(0.2, 0.25, 0.25, 0.3)

share_by_reason_dbl

Share by reason (a double vector), Default: c(0.15, 0.06, 0.01, 0.06, 0.68)

share_by_referrer_dbl

Share by referrer (a double vector), Default: c(rep(0.1, 3), rep(0.06, 2), 0.1, 0.08, 0.1, 0.3)

share_by_role_dbl

Share by role (a double vector), Default: c(0.12, 0.65, 0.01, 0.2, 0.02)

share_by_sex_dbl

Share by sex (a double vector), Default: c(0.05, 0.55, 0.4)

share_is_para_1L_dbl

Share is para (a double vector of length one), Default: 0.28

sheets_int

Sheets (an integer vector), Default: c(1, 3:8)

sports_1L_int

Sports (an integer vector of length one), Default: integer(0)

sports_tab_1L_int

Sports tab (an integer vector of length one), Default: 2

start_date_1L_chr

Start date (a character vector of length one), Default: '2022-03-01'

uid_vars_chr

Unique identifier variables (a character vector), Default: c("Client ID", "UID")

unit_cost_lup

Unit cost (a lookup table), Default: NULL

Value

Clients (a tibble)