Make service use variables

make_service_use_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make service use variables. The function returns Service use (a character vector).

Usage

make_service_use_vars(
  X_Ready4useDyad,
  active_base_1L_chr = "Active",
  patterns_ls = list(c("[[:space:]]", "")),
  prefix_1L_chr = "Cumulative",
  separation_after_dbl = c(3, 6),
  service_var_1L_chr = "Service",
  tenure_var_1L_chr = "Tenure"
)

Arguments

X_Ready4useDyad

PARAM_DESCRIPTION

active_base_1L_chr

Active base (a character vector of length one), Default: 'Active'

patterns_ls

Patterns (a list), Default: list(c("[:space:]", ""))

prefix_1L_chr

Prefix (a character vector of length one), Default: 'Cumulative'

separation_after_dbl

Separation after (a double vector), Default: c(3, 6)

service_var_1L_chr

Service variable (a character vector of length one), Default: 'Service'

tenure_var_1L_chr

Tenure variable (a character vector of length one), Default: 'Tenure'

Value

Service use (a character vector)