Update ingested data

update_ingested_data() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update ingested data. The function returns Datasets (a list).

Usage

update_ingested_data(
  datasets_ls,
  categories_chr = c("Individual Sports", "Aesthetic Sports", "Winter Sports"),
  exclude_chr = c("Cost", "Duration"),
  imputed_uid_pfx_chr = "UNK",
  missing_1L_chr = "0",
  provider_id_1L_chr = "ProviderID",
  provider_location_1L_chr = "ProviderState",
  uid_vars_chr = c("MedlinksID", "AISID")
)

Arguments

datasets_ls

Datasets (a list)

categories_chr

Categories (a character vector), Default: c("Individual Sports", "Aesthetic Sports", "Winter Sports")

exclude_chr

Exclude (a character vector), Default: c("Cost", "Duration")

imputed_uid_pfx_chr

Imputed unique identifier prefix (a character vector), Default: 'UNK'

missing_1L_chr

Missing (a character vector of length one), Default: '0'

provider_id_1L_chr

Provider identity (a character vector of length one), Default: 'ProviderID'

provider_location_1L_chr

Provider location (a character vector of length one), Default: 'ProviderState'

uid_vars_chr

Unique identifier variables (a character vector), Default: c("MedlinksID", "AISID")

Value

Datasets (a list)